backend/requirements.txt
fastapi
uvicorn
web3
pydantic
requests
web3
python-dotenv